Add architecture explorer tool route
