+++ ui/partner-hub/app/(routes)/tools/architecture-explorer/page.tsx
@@ -1,0 +1,5 @@
+import { ArchitectureExplorer } from "@/components/architecture-explorer/architecture-explorer";
+
+export default function ArchitectureExplorerPage() {
+  return <ArchitectureExplorer />;
+}
