+++ ui/partner-hub/components/architecture-explorer/diagramScenarios.ts
